requirements.txt
streamlit==1.28.1
langchain==0.0.350
openai==1.3.7
faiss-cpu==1.7.4
python-dotenv==1.0.0
PyMuPDF==1.23.8
langchain-openai==0.0.2
langchain-community==0.0.10
groq==0.4.2
scikit-learn==1.3.0
numpy==1.24.3
pypdf2==3.0.1 